backend/pyproject.toml
[project]
name = "munch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asgiref>=3.11.0",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "sqlparse>=0.5.3",
]
